more floats -.-

+++ stream.c
@@ -71,7 +71,7 @@ void stream_calc_throughput(stream *s, float alpha) {
   if (s == NULL) return;
   if (s->cur_request == NULL) return;
   float t_new = s->cur_request->throughput;
-  int t_cur = (alpha * t_new) + ((1 - alpha) * s->throughput);
+  float t_cur = (alpha * t_new) + ((1 - alpha) * s->throughput);
   s->throughput = t_cur;
 }
